+++ extensions/default-web-browser-manager/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Default Web Browser Manager Changelog
 
+## [Add Open Default Browser Command] - {PR_MERGE_DATE}
+
+- Add "Open Default Browser" command to quickly open the current default browser
+
 ## [Rename Command] - 2025-12-17
 
 - Rename "Choose" to "Set" so it better aligns with the default "System" setting commands